Convert the snippet chooser to use static onload handlers
