--- wagtail/snippets/templates/wagtailsnippets/chooser/chosen.js
@@ -1,4 +1,0 @@
-function(modal, jsonData) {
-    modal.respond('snippetChosen', jsonData['result']);
-    modal.close();
-}
